# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Vietnamese stock market prediction system — ML algorithms (VWMA, EMA, LSTM, ARIMA-GARCH, Ensemble) for VN30 stocks, gold, NASDAQ, crypto, and fuel prices.",
        "title": "Go Stock Prediction API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8118",
    "basePath": "/",
    "paths": {
...
        "/api/crypto/chart": {
            "get": {
                "description": "Returns chronologically ordered date labels and closing prices for charting",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crypto"
                ],
                "summary": "Get crypto price chart data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coin ID (e.g. bitcoin, ethereum)",
                        "name": "coin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of days to look back (default 30)",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.cryptoChartResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/api/crypto/latest": {
            "get": {
                "description": "Returns the most recent price, market cap, and 24h volume for each tracked cryptocurrency",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crypto"
                ],
                "summary": "Get latest crypto prices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.cryptoLatestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/api/crypto/predictions": {
            "get": {
                "description": "Returns paginated crypto predictions with optional filtering by coin, algorithm, and status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crypto Predictions"
                ],
                "summary": "List crypto predictions with pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by coin ID (e.g. bitcoin)",
                        "name": "coin",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by algorithm name",
                        "name": "algorithm",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending, confirmed, wrong)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction (asc, desc)",
                        "name": "sort_dir",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page, max 100 (default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.cryptoPredictionsPageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/api/crypto/predictions/chart": {
            "get": {
                "description": "Returns chronologically ordered predicted vs actual price data for charting, optionally filtered by algorithm",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crypto Predictions"
                ],
                "summary": "Get crypto prediction chart data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coin ID (e.g. bitcoin, ethereum)",
                        "name": "coin",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Algorithm name (e.g. lstm_nn)",
                        "name": "algorithm",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of days to look back (default 30)",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.cryptoPredictionChartResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/api/crypto/predictions/latest": {
            "get": {
                "description": "Returns the most recent prediction per (coin_id, algorithm) combination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crypto Predictions"
                ],
                "summary": "Get latest crypto predictions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.cryptoPredictionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseFailure"
                        }
                    }
                }
            }
        },
        "/api/crypto/prices": {
            "get": {
                "description": "Returns close price, market cap, and 24h volume for a given coin over the specified number of days",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crypto"
                ],
                "summary": "Get historical crypto prices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coin ID (e.g. bitcoin, ethereum)",
                        "name": "coin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of days to look back (default 30)",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.cryptoPricesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseFailure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.ResponseFailure": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
        "server.cryptoChartResponse": {
            "type": "object",
            "properties": {
                "coin_id": {
                    "type": "string"
                },
                "dates": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "prices": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
        "server.cryptoLatestItem": {
            "type": "object",
            "properties": {
                "close_price": {
                    "type": "number"
                },
                "coin_id": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "market_cap": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "trading_date": {
                    "type": "string"
                },
                "volume_24h": {
                    "type": "number"
                }
            }
        },
        "server.cryptoLatestResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.cryptoLatestItem"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "server.cryptoPredictionChartPoint": {
            "type": "object",
            "properties": {
                "actual_price": {
                    "type": "number"
                },
                "confidence": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "predicted_price": {
                    "type": "number"
                }
            }
        },
        "server.cryptoPredictionChartResponse": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "coin_id": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.cryptoPredictionChartPoint"
                    }
                }
            }
        },
        "server.cryptoPredictionItem": {
            "type": "object",
            "properties": {
                "accuracy": {
                    "type": "number"
                },
                "actual_price": {
                    "type": "number"
                },
                "algorithm_name": {
                    "type": "string"
                },
                "coin_id": {
                    "type": "string"
                },
                "confidence": {
                    "type": "number"
                },
                "current_price": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "predicted_price": {
                    "type": "number"
                },
                "prediction_date": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "target_date": {
                    "type": "string"
                }
            }
        },
        "server.cryptoPredictionsPageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.cryptoPredictionItem"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "server.cryptoPredictionsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.cryptoPredictionItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "server.cryptoPricePoint": {
            "type": "object",
            "properties": {
                "close_price": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "market_cap": {
                    "type": "number"
                },
                "volume_24h": {
                    "type": "number"
                }
            }
        },
        "server.cryptoPricesResponse": {
            "type": "object",
            "properties": {
                "coin_id": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.cryptoPricePoint"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token from POST /api/auth/login. Format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}